+++ react-frontend/src/components/SideNavigation.jsx
