+++ react-frontend/src/components/formfield.jsx
@@ -1,0 +1,17 @@
+import { FormControl, FormField, FormItem, FormLabel } from "./ui/form";
+import { Input } from "./ui/input";
+
+export const SimpleFormField = ({ control, name, label, type = "text" }) => (
+    <FormField
+      control={control}
+      name={name}
+      render={({ field }) => (
+        <FormItem>
+          <FormLabel className=" font-normal">{label}</FormLabel>
+          <FormControl>
+            <Input type={type} {...field} />
+          </FormControl>
+        </FormItem>
+      )}
+    />
+  );
