+++ react-frontend/src/features/departments/index.jsx
@@ -9,67 +9,13 @@ import {
 } from "@/components/ui/dialog";
 import {
   Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
 } from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import {
-  Table,
-  TableBody,
-  TableCaption,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "@/components/ui/table";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
+import { DepartmentsTable } from "./list";
+import { SimpleFormField } from "@/components/formfield";
 
-const DepartmentsTable = ({ depts }) => (
-  <Table>
-    <TableCaption>List of departments</TableCaption>
-    <TableHeader>
-      <TableRow>
-        <TableHead>S.no</TableHead>
-        <TableHead>Department name</TableHead>
-        <TableHead>HOD</TableHead>
-        <TableHead>Faculty</TableHead>
-        <TableHead>Staff</TableHead>
-        <TableHead>Courses</TableHead>
-      </TableRow>
-    </TableHeader>
-    <TableBody>
-      {depts.map((dept, index) => (
-        <TableRow key={dept.id}>
-          <TableCell>{index+1}</TableCell>
-          <TableCell>{dept.name}</TableCell>
-          <TableCell>{dept.hod}</TableCell>
-          <TableCell>{dept.faculty}</TableCell>
-          <TableCell>{dept.staff}</TableCell>
-          <TableCell>{dept.courses}</TableCell>
-        </TableRow>
-      ))}
-    </TableBody>
-  </Table>
-);
-const AdmiFormField = ({ control, name, label, type = "text" }) => (
-  <FormField
-    control={control}
-    name={name}
-    render={({ field }) => (
-      <FormItem>
-        <FormLabel className=" font-normal">{label}</FormLabel>
-        <FormControl>
-          <Input type={type} {...field} />
-        </FormControl>
-      </FormItem>
-    )}
-  />
-);
-
-export function PrincipalDeptsPage() {
+export function DepartmentsPage() {
   const [departments, setDepartments] = useState([]);
   const [newDeptDialog, setnewDeptDialog] = useState(false);
   const { control, ...rest } = useForm();
@@ -82,7 +28,7 @@ export function PrincipalDeptsPage() {
 
   return (
     <>
-      <Button variant="ghost" onClick={(e) => setnewDeptDialog(true)}>
+      <Button variant="ghost" onClick={() => setnewDeptDialog(true)}>
         New department
       </Button>
       <DepartmentsTable depts={departments} />
@@ -97,12 +43,12 @@ export function PrincipalDeptsPage() {
           </DialogHeader>
           <Form control={control} {...rest}>
             <form method="post" onSubmit={rest.handleSubmit(onNewCourse)} className="flex gap-3 flex-col">
-              <AdmiFormField
+              <SimpleFormField
                 control={control}
                 name="name"
                 label="Department name"
               />
-              <AdmiFormField
+              <SimpleFormField
                 control={control}
                 name="hod"
                 label="Head of the department"
